refactor ratings functionality
